feat(desktop): project thread displays on the owning instance
