+++ apps/desktop/src/main/__tests__/federation-runtime.test.ts
@@ -380,14 +380,18 @@ describe("DesktopFederationRuntime", () => {
     owner.forwardLocalBackendEvent(pricing);
     owner.forwardLocalBackendEvent(pricing);
     expect(frames.at(-1)).toMatchObject({ params: { accountingPatch: { changes: [] } } });
-    expect(published.at(-1)?.notification).toEqual(pricing.notification);
+    expect(published.at(-1)?.notification).toMatchObject({ method: "thread/pricing/updated", params: {
+      threadId: "thread-1", displayInvalidated: true, pricing: { lines: [], summaries: [] },
+    } });
+    expect(JSON.stringify(frames.at(-1))).not.toContain("compaction-99");
     const subAgents: AgentEvent = { backend: "codex", notification: { method: "thread/subAgents/updated", params: {
       threadId: "thread-1", subAgents: [{ monitorId: "monitor", task: "History ".repeat(1000), status: "success", createdAt: 1, updatedAt: 1 }],
     } } };
     owner.forwardLocalBackendEvent(subAgents);
     owner.forwardLocalBackendEvent(subAgents);
-    expect(frames.at(-1)).toMatchObject({ params: { accountingPatch: { changes: [] } } });
-    expect(published.at(-1)?.notification).toEqual(subAgents.notification);
+    expect(published.at(-1)?.notification).toEqual({ method: "thread/subAgents/updated", params: { threadId: "thread-1" } });
+    expect(Buffer.byteLength(JSON.stringify(frames.at(-1)))).toBeLessThan(1000);
+    expect(JSON.stringify(frames.at(-1))).not.toContain("History");
     // A reconnect needs catch-up even when no later event arrives to reveal
     // a missed request-user-input. Replaying the subscription supplies it.
     if (viaGateway) gateway.replayRelayedEventSubscriptions("owner_one");
