+++ packages/shared/src/contracts/normalized-app-server.ts
@@ -1,3 +1,7 @@
+import type { SubAgentLens } from "../subagent-kind";
+import type { ToolAccountingTotals } from "../thread-tool-display";
+import type { ThreadIncidentSummary } from "../thread-incident-summary";
+import type { ThreadPricingDisplay } from "../thread-pricing-display";
 import type { AutomationRunOutputDecision } from "./automations";
 import type { ScheduledThreadAction } from "./scheduled-thread-actions";
 import type {
@@ -1009,7 +1013,35 @@ export type RenameThreadResponse = {
   renamedAt: number;
 };
 
+/** Display resources have independent pages. Transcript loads never carry accounting ledgers. */
+export type ThreadDisplayRead = {
+  subAgentLens?: SubAgentLens;
+  resource: "transcript" | "accounting" | "pricing" | "tools" | "incident" | "subagents" | "subagent";
+  firstWarningAt?: number;
+  largeOutputThresholdChars?: number;
+  monitorId?: string;
+  cursor?: string;
+  limit?: number;
+  /** Loaded transcript turns whose finalized usage activities need refreshing. */
+  turns?: AppServerThreadTurnMetadata[];
+};
+
+export type ThreadDisplayData = {
+  subAgentLens?: SubAgentLens;
+  subAgentCounts?: Record<SubAgentLens, number>;
+  subAgents?: ThreadSubAgentSummary[];
+  subAgent?: ThreadSubAgentSummary;
+  incident?: ThreadIncidentSummary;
+  pricing: Omit<ThreadPricingDisplay, "rows" | "spendByModel">;
+  pricingPage?: ThreadPricingDisplay;
+  toolsPage?: ThreadToolAccounting;
+  toolTotals?: ToolAccountingTotals;
+  nextCursor?: string;
+  revision: string;
+};
+
 export type AppServerReadThreadRequest = {
+  display?: ThreadDisplayRead;
   /** Bounded diagnostic attribution; never transcript content. */
   readReason?: "thread-view" | "star-map-card";
   /** Opt into conditional page revalidation. Empty string requests a first revision. */
@@ -1035,6 +1067,7 @@ export type AppServerReadThreadRequest = {
 };
 
 export type AppServerReadThreadResponse = {
+  display?: ThreadDisplayData;
   /** Opaque hash of the complete owner response, excluding observation timings. */
   replayRevision?: string;
   /** Only in response to knownRevision. Reuse that exact cached response, not this empty replay. */
@@ -1747,6 +1780,7 @@ export type AppServerNotification =
   | {
       method: "thread/pricing/updated";
       params: {
+        displayInvalidated?: true;
         threadId: string;
         pricing: {
           /** Observed context compactions, oldest first. */
@@ -1760,6 +1794,8 @@ export type AppServerNotification =
   | {
       method: "thread/toolAccounting/updated";
       params: {
+        incidentSummary?: ThreadIncidentSummary;
+        displayInvalidated?: true;
         threadId: string;
         toolAccounting: ThreadToolAccounting;
         /**
