+++ packages/shared/src/subagent-kind.ts
@@ -1,0 +1,93 @@
+import type { ThreadSubAgentSummary } from "./contracts/navigation";
+
+export type SubAgentLens = "harness" | "token-miser" | "pwragent";
+
+export function isCodexNativeSubAgent(subAgent: ThreadSubAgentSummary): boolean {
+  return subAgent.monitorId.startsWith("codex-native:");
+}
+
+export function isSystemTitleHelperSubAgent(
+  subAgent: ThreadSubAgentSummary,
+): boolean {
+  return subAgent.monitorId.startsWith("system:title-helper:");
+}
+
+export function isTokenMiserSubAgent(
+  subAgent: ThreadSubAgentSummary,
+): boolean {
+  return subAgent.monitorId.startsWith("system:token-miser:");
+}
+
+/**
+ * Groups sub-agents by the system that owns their lifecycle. The harness lens
+ * is intentionally provider-neutral even though Codex is the only harness
+ * that projects native workers today.
+ */
+export function subAgentLens(subAgent: ThreadSubAgentSummary): SubAgentLens {
+  if (isTokenMiserSubAgent(subAgent)) {
+    return "token-miser";
+  }
+  if (isCodexNativeSubAgent(subAgent)) {
+    return "harness";
+  }
+  return "pwragent";
+}
+
+export function subAgentOriginLabel(
+  subAgent: ThreadSubAgentSummary,
+): string | undefined {
+  if (isSystemTitleHelperSubAgent(subAgent)) {
+    return "PwrAgent system helper";
+  }
+  if (isTokenMiserSubAgent(subAgent)) {
+    return "PwrAgent Token Miser gate";
+  }
+  if (isCodexNativeSubAgent(subAgent)) {
+    return "Codex native spawnAgent";
+  }
+  if (subAgent.monitorId.startsWith("review:")) {
+    return "PwrAgent code review";
+  }
+  return "PwrAgent task monitor";
+}
+
+export function subAgentOriginSentence(
+  subAgent: ThreadSubAgentSummary,
+): string | undefined {
+  const label = subAgentOriginLabel(subAgent);
+  return label ? `Spawned by ${label}.` : undefined;
+}
+
+export function subAgentUsageLabel(subAgent: ThreadSubAgentSummary): string {
+  if (isSystemTitleHelperSubAgent(subAgent)) {
+    return "System";
+  }
+  if (isTokenMiserSubAgent(subAgent)) {
+    return "Gate";
+  }
+  if (isCodexNativeSubAgent(subAgent)) {
+    return "Codex";
+  }
+  if (subAgent.monitorId.startsWith("review:")) {
+    return "Review";
+  }
+  return "Monitor";
+}
+
+export function subAgentPricingUsageTitle(
+  subAgent: ThreadSubAgentSummary,
+): string {
+  if (isSystemTitleHelperSubAgent(subAgent)) {
+    return "Thread naming";
+  }
+  if (isTokenMiserSubAgent(subAgent)) {
+    return "Token Miser gate";
+  }
+  if (isCodexNativeSubAgent(subAgent)) {
+    return "Codex sub-agent usage";
+  }
+  if (subAgent.monitorId.startsWith("review:")) {
+    return "Review usage";
+  }
+  return "Monitor usage";
+}
